# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nulls Report production portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/logout:
    post:
      operationId: logoutPortalSession
      tags: [account]
      summary: End the current session
      responses:
        "204":
          description: Signed out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/methods/{provider}/disconnect:
    post:
      operationId: disconnectAuthMethod
      tags: [account]
      summary: Disconnect a linked authentication method
      description: >
        Removes the provider link from the account. The account itself is
        never deleted. The client warns when this is the only method.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [discord, nulls_connect]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/discord/start:
    get:
      operationId: discordOauthStart
      tags: [account]
      summary: Start Discord OAuth (browser redirect)
      parameters:
        - name: returnTo
          in: query
          required: false
          schema: { type: string }
      responses:
        "302":
          description: Redirect to Discord
  /me:
    get:
      operationId: getCurrentUser
      tags: [account]
      summary: Get the signed-in portal user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateCurrentUser
      tags: [account]
      summary: Update the signed-in portal user's account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: number }
        tag: { type: ["string", "null"] }
        displayName: { type: string }
        role:
          type: string
          enum: [user, moderator, administrator]
        blocked: { type: boolean }
        nullsConnectId: { type: ["string", "null"] }
        nullsConnectName: { type: ["string", "null"] }
        avatarPath: { type: ["string", "null"] }
        discordId: { type: ["string", "null"] }
        discordUsername: { type: ["string", "null"] }
        preferences:
          type: object
          additionalProperties: true
        authMethods:
          type: array
          items: { $ref: "#/components/schemas/AuthMethod" }
        hasTrustedAuth: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, tag, displayName, role, preferences, authMethods, hasTrustedAuth, createdAt]
    AuthMethod:
      type: object
      properties:
        provider: { type: string, enum: [discord, nulls_connect] }
        name: { type: string }
        label: { type: ["string", "null"] }
        linkedAt: { type: string, format: date-time }
      required: [provider, name, label, linkedAt]
...
    UpdateUserInput:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        preferences:
          type: object
          additionalProperties: true
      description: At least one field is required.